+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-- Added `trial_05B_wikihow.py` to test on WikiHow dataset
-- Added `test_wikihow_dataset.py` to produce cleaned and truncated version of WikiHow dataset
-- Added `wikihow_data/cleaned_10000.csv` a cleaned version of 10K samples from original dataset.
+- Added summaries for `trial_05B_wikihow.py`
+